backups/containers/backup_20251111_204109/engineer_b_api_app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
psycopg2-binary==2.9.9
httpx
google-genai
